# Backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Dokkan Backend API",
    "description": "High-coverage API documentation for the Dokkan e-commerce platform, including auth, stores, catalog, cart, orders, and reviews."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List all products",
        "description": "Returns a paginated list of products with optional status and sort filters.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "example": 1,
            "description": "Page number",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "example": 20,
            "description": "Items per page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "example": "Active",
            "description": "Filter by product status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "example": "createdAt",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "required": false,
            "example": "desc",
            "description": "Sort direction",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product (Store Owner only)",
        "description": "Creates a new product. Use multipart/form-data to upload an image file under field `image`.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "storeId",
                  "subCategoryId",
                  "title",
                  "price"
                ],
                "properties": {
                  "storeId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "1b3b0de0-b3f7-4d17-9df1-c1b3d31a3fd0"
                  },
                  "subCategoryId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "2c4a1e18-38f5-4d17-b8a0-0a3d4c15cf66"
                  },
                  "title": {
                    "type": "string",
                    "example": "Summer Tee"
                  },
                  "description": {
                    "type": "string",
                    "example": "Comfortable cotton t-shirt"
                  },
                  "price": {
                    "type": "number",
                    "example": 29.99
                  },
                  "stockQuantity": {
                    "type": "integer",
                    "example": 100
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ],
                    "example": "Active"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image file"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stores/": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Create a new store",
        "description": "Creates a new store for the authenticated user. Store owner role will be applied.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Store created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "type": "object",
                    "required": [
                      "name",
                      "subdomain",
                      "logoUrl"
                    ],
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "My Store"
                      },
                      "subdomain": {
                        "type": "string",
                        "example": "my-store"
                      },
                      "logoUrl": {
                        "type": "string",
                        "example": "https://cdn.example.com/logo.png"
                      },
                      "description": {
                        "type": "string",
                        "example": "A store selling gadgets"
                      },
                      "coverBannerUrl": {
                        "type": "string",
                        "example": "https://cdn.example.com/banner.png"
                      },
                      "businessAddress": {
                        "type": "string",
                        "example": "123 Main St, Cairo"
                      },
                      "vatNumber": {
                        "type": "string",
                        "example": "VAT-123456"
                      },
                      "supportEmail": {
                        "type": "string",
                        "format": "email",
                        "example": "support@example.com"
                      },
                      "phoneNumber": {
                        "type": "string",
                        "example": "+201234567890"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "Browse active stores with pagination",
        "description": "List stores with optional pagination, status filter, and sorting.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "example": 20,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "example": "Active",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "example": "createdAt",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "required": false,
            "example": "desc",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stores retrieved successfully"
          }
        }
      }
    },
...
    "/api/stores/stores": {
      "delete": {
        "tags": [
          "Stores"
        ],
        "summary": "Create a new store",
        "description": "Creates a new store for the authenticated user. Store owner role will be applied.",
        "responses": {
          "201": {
            "description": "Store created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "type": "object",
                    "required": [
                      "name",
                      "subdomain",
                      "logoUrl"
                    ],
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "My Store"
                      },
                      "subdomain": {
                        "type": "string",
                        "example": "my-store"
                      },
                      "logoUrl": {
                        "type": "string",
                        "example": "https://cdn.example.com/logo.png"
                      },
                      "description": {
                        "type": "string",
                        "example": "A store selling gadgets"
                      },
                      "coverBannerUrl": {
                        "type": "string",
                        "example": "https://cdn.example.com/banner.png"
                      },
                      "businessAddress": {
                        "type": "string",
                        "example": "123 Main St, Cairo"
                      },
                      "vatNumber": {
                        "type": "string",
                        "example": "VAT-123456"
                      },
                      "supportEmail": {
                        "type": "string",
                        "format": "email",
                        "example": "support@example.com"
                      },
                      "phoneNumber": {
                        "type": "string",
                        "example": "+201234567890"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}